style(Tabs): change tab label style
